refactor: reorganize service interface folders
