+++ backend/src/modules/auth/auth.module.ts
@@ -4,11 +4,11 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { RefreshToken } from 'src/database/entities/refresh-token.entity';
-import { AUTH_SERVICE } from '../../common/interfaces/auth-service.interface';
 import { UsersModule } from '../users/users.module';
 import { VerificationModule } from '../verification/verification.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
+import { AUTH_SERVICE } from './interfaces/auth-service.interface';
 import { JwtRefreshStrategy } from './strategies/jwt-refresh.strategy';
 import { JwtResetStrategy } from './strategies/jwt-reset.strategy';
 import { JwtStrategy } from './strategies/jwt.strategy';
