+++ backend/src/modules/auth/interfaces/auth-service.interface.ts
@@ -1,8 +1,8 @@
 import type { User } from 'src/database/entities/user.entity';
-import type { LoginResponseDto } from '../../modules/auth/dto/login-response.dto';
-import type { LoginDto } from '../../modules/auth/dto/login.dto';
-import type { RegisterResponseDto } from '../../modules/auth/dto/register-response.dto';
-import type { RegisterDto } from '../../modules/auth/dto/register.dto';
+import type { LoginResponseDto } from '../dto/login-response.dto';
+import type { LoginDto } from '../dto/login.dto';
+import type { RegisterResponseDto } from '../dto/register-response.dto';
+import type { RegisterDto } from '../dto/register.dto';
 
 /**
  * Service token for dependency injection of authentication service.
